VPS deadling improoved

+++ database/migrations/2025_01_11_120601_create_processes_table.php
@@ -42,8 +42,8 @@ return new class extends Migration
                 ->on('process_responsible_people');
 
             // auto
-            $table->smallInteger('order_priority') // used in deadline status
-                ->default(0);
+            $table->decimal('order_priority', 8, 3) // used in deadline status
+                ->default(0.000);
 
             // Stage 1 (ВП)
             // required and immutable after stage 1
